downloader_general/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
curl-cffi==0.13.0
distro==1.9.0
frozendict==2.4.7
gitdb==4.0.12
gitpython==3.1.46
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
multitasking==0.0.12
numpy==2.4.2
openai==2.25.0
pandas==3.0.1
platformdirs==4.9.4
polars==1.38.1
polars-runtime-32==1.38.1
portalocker==3.2.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
pyyaml==6.0.3
qdrant-client==1.17.0
regex==2026.2.28
requests==2.32.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.47
tabulate==0.9.0
tiktoken==0.12.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wbgapi==1.0.14
websockets==16.0
yfinance==1.2.0
